pyproject.toml
[project]
name = "lidar-features"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "evalio",
    "gdown>=5.2.0",
    "ipython>=8.32.0",
    "loam",
    "matplotlib>=3.10.0",
    "polars>=1.21.0",
    "pyqt6>=6.8.0",
    "pyserde[yaml]>=0.23.0",
    "rerun-sdk>=0.21.0",
    "seaborn>=0.13.2",
    "tabulate>=0.9.0",
    "tqdm>=4.67.1",
]

[tool.uv.sources]
evalio = { workspace = true }
loam = { workspace = true }

[tool.uv.workspace]
members = ["evalio", "loam"]

[dependency-groups]
dev = [
    "compdb>=0.2.0",
    "pybind11-stubgen>=2.5.1",
]
